requirements.txt
langchain==0.1.13
langchain_openai==0.1.0
python-dotenv==1.0.1
